improved logging, handle more errors, convert more layout sets

+++ cli-tools/altinn-app-cli/Fev3Tov4/LayoutSetRewriter/LayoutSetUpgrader.cs
@@ -12,6 +12,7 @@ class LayoutSetUpgrader
     private readonly string layoutSetName;
     private readonly string applicationMetadataFile;
     private JsonObject? layoutSetsJson = null;
+    private JsonObject? layoutSettingsJson = null;
 
     public LayoutSetUpgrader(string uiFolder, string layoutSetName, string applicationMetadataFile)
     {
@@ -29,7 +30,7 @@ class LayoutSetUpgrader
     {
         // Read applicationmetadata.json file
         var appMetaText = File.ReadAllText(applicationMetadataFile);
-        var appMetaJson = JsonNode.Parse(appMetaText);
+        var appMetaJson = JsonNode.Parse(appMetaText, null, new JsonDocumentOptions() { CommentHandling = JsonCommentHandling.Skip, AllowTrailingCommas = true });
         if (appMetaJson is not JsonObject appMetaJsonObject)
         {
             warnings.Add($"Unable to parse {applicationMetadataFile}, skipping layout sets upgrade");
@@ -108,8 +109,25 @@ class LayoutSetUpgrader
             return;
         }
 
-        var jsonString = $@"{{""$schema"": ""https://altinncdn.no/schemas/json/layout/layout-sets.schema.v1.json"", ""sets"": [{{""id"": ""{layoutSetName}"", ""dataType"": ""{dataTypeId}"", ""tasks"": [""{taskId}""]}}]}}";
-        layoutSetsJson = JsonNode.Parse(jsonString)?.AsObject();
+        var layoutSetsJsonString = $@"{{""$schema"": ""https://altinncdn.no/schemas/json/layout/layout-sets.schema.v1.json"", ""sets"": [{{""id"": ""{layoutSetName}"", ""dataType"": ""{dataTypeId}"", ""tasks"": [""{taskId}""]}}]}}";
+        layoutSetsJson = JsonNode.Parse(layoutSetsJsonString)?.AsObject();
+
+        // Generate basic Settings.json file
+        List<string>? order = null;
+        if (Directory.Exists(Path.Combine(uiFolder, "layouts")))
+        {
+            order = Directory.GetFiles(Path.Combine(uiFolder, "layouts"), "*.json").Select(f => $@"""{Path.GetFileNameWithoutExtension(f)}""").ToList();
+        }
+        else if (File.Exists(Path.Combine(uiFolder, "FormLayout.json")))
+        {
+            order = new List<string>() { @"""FormLayout""" };
+        }
+
+        if (order != null)
+        {
+            var layoutSettingsJsonString = $@"{{""$schema"": ""https://altinncdn.no/schemas/json/layout/layoutSettings.schema.v1.json"", ""pages"": {{""order"": [{string.Join(", ", order)}]}}}}";
+            layoutSettingsJson = JsonNode.Parse(layoutSettingsJsonString)?.AsObject();
+        }
     }
 
     public async Task Write()
@@ -119,17 +137,47 @@ class LayoutSetUpgrader
             return;
         }
 
+        JsonSerializerOptions options = new JsonSerializerOptions
+        {
+            WriteIndented = true,
+            Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping,
+        };
+
         // Create new layout set folder
         Directory.CreateDirectory(Path.Combine(uiFolder, layoutSetName));
 
         // Move existing files to new layout set
         var oldLayoutsPath = Path.Combine(uiFolder, "layouts");
         var newLayoutsPath = Path.Combine(uiFolder, layoutSetName, "layouts");
-        Directory.Move(oldLayoutsPath, newLayoutsPath);
+        if (Directory.Exists(oldLayoutsPath))
+        {
+            Directory.Move(oldLayoutsPath, newLayoutsPath);
+        }
+        else if (File.Exists(Path.Combine(uiFolder, "FormLayout.json")))
+        {
+            Directory.CreateDirectory(newLayoutsPath);
+            File.Move(Path.Combine(uiFolder, "FormLayout.json"), Path.Combine(newLayoutsPath, "FormLayout.json"));
+        }
+        else
+        {
+            warnings.Add($"Unable to find any layout files in {uiFolder}");
+        }
 
         var oldSettingsPath = Path.Combine(uiFolder, "Settings.json");
         var newSettingsPath = Path.Combine(uiFolder, layoutSetName, "Settings.json");
-        File.Move(oldSettingsPath, newSettingsPath);
+        if (File.Exists(oldSettingsPath))
+        {
+            File.Move(oldSettingsPath, newSettingsPath);
+        }
+        else if (layoutSettingsJson != null)
+        {
+            // Write new Settings.json
+            await File.WriteAllTextAsync(newSettingsPath, layoutSettingsJson.ToJsonString(options));
+        }
+        else
+        {
+            warnings.Add($"Unable to find Settings.json in {uiFolder}, also unable to find layout files to generate one");
+        }
 
         var oldRuleConfigurationPath = Path.Combine(uiFolder, "RuleConfiguration.json");
         var newRuleConfigurationPath = Path.Combine(uiFolder, layoutSetName, "RuleConfiguration.json");
@@ -146,11 +194,6 @@ class LayoutSetUpgrader
         }
 
         // Write new layout-sets.json
-        JsonSerializerOptions options = new JsonSerializerOptions
-        {
-            WriteIndented = true,
-            Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping,
-        };
         await File.WriteAllTextAsync(Path.Combine(uiFolder, "layout-sets.json"), layoutSetsJson.ToJsonString(options));
     }
 }
